app/requirements.txt
fastapi
playwright
pymongo
uvicorn
celery
flower
redis
crawl4ai
asgiref
langchain
langchain-ollama
celery-redbeat
streamlit